# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation for LIWACH system",
        "description": "API Documentation for LIWACH system",
        "contact": {
            "email": "edenabdisa90@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://liwachapi.herokuapp.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of user",
                "description": "Returns list of user",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store new User",
                "description": "Returns User data",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/user/{id}": {            
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update existing user",
                "description": "Returns updated user data",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete existing user",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "big integer",
                        "format": "int64",
                        "example": 1
                    },
                    "first_name": {
                        "title": "Last Name",
                        "description": "The first Name",
                        "type": "string",
                        "example": "Eden "
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "The last Name",
                        "type": "string",
                        "example": "Abdisa "
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "format": "eden@yahoo.com"
                    },
                    "profile_picture": {
                        "title": "Picture file name",
                        "description": "This name will be concated with the image collection folder",
                        "type": "string",
                        "example": "myprofile.jpg"
                    },
                    "phone_number": {
                        "title": "Phone Number",
                        "description": "Phone number including country/city code",
                        "type": "string", 
                        "example": "+2519443322"
                    },
                    "TIN_picture": {
                        "title": "Picture file name",
                        "description": "This name will be concated with the image collection folder and the tin number will be viewed by the admin to approve the user",
                        "type": "string",
                        "example": "mytin.jpg"
                    },
                    "status": {
                        "title": "Status",
                        "description": "The user is active at first for user and blocked for company",
                        "type": "string",
                        "example": "active or blocked"
                    },
                    "birthdate": {
                        "title": "Birthdate",
                        "description": "The birthdate of the user",
                        "type": "string",
                        "format": "date",
                        "example":"2020-01-27"
                    },
                    "type": {
                        "title": "Type of user",
                        "description": "This identify the account whether it is an organizations or a user or admin",
                        "type": "string",
                        "example": "user or admin or company"
                    },
                    
                    "address_id": {
                        "title": "Address ID",
                        "description": "The ID of the address",
                        "type": "big integer",
                        "format": 99
                    },
                    "membership_id": {
                        "title": "Membership ID ",
                        "description": "This is the ID of the membership that the user or organization choose",
                        "type": "big integer",
                        "example": 8
                    },
                    "email_verified_at": {
                        "title": "Email verified at",
                        "description": "Email verified at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "rememberToken": {
                        "title": "Login Token",
                        "description": "The token optained when user logged in",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "timestamp": {
                        "title": "Timestamp",
                        "description": "This is the timestamp that we got when this data is added.",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                }
            },
            "UserRequest": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "first_name": {
                        "title": "Last Name",
                        "description": "The first Name",
                        "type": "string",
                        "example": "Eden "
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "The last Name",
                        "type": "string",
                        "example": "Abdisa "
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "format": "eden@yahoo.com"
                    },
                    "profile_picture": {
                        "title": "Picture file name",
                        "description": "This name will be concated with the image collection folder",
                        "type": "string",
                        "example": "myprofile.jpg"
                    },
                    "phone_number": {
                        "title": "Phone Number",
                        "description": "Phone number including country/city code",
                        "type": "string", 
                        "example": "+2519443322"
                    },
                    "TIN_picture": {
                        "title": "Picture file name",
                        "description": "This name will be concated with the image collection folder and the tin number will be viewed by the admin to approve the user",
                        "type": "string",
                        "example": "mytin.jpg"
                    },
                    "status": {
                        "title": "Status",
                        "description": "The user is active at first for user and blocked for company",
                        "type": "string",
                        "example": "active or blocked"
                    },
                    "birthdate": {
                        "title": "Birthdate",
                        "description": "The birthdate of the user",
                        "type": "string",
                        "format": "date",
                        "example":"2020-01-27"
                    },
                    "type": {
                        "title": "Type of user",
                        "description": "This identify the account whether it is an organizations or a user or admin",
                        "type": "string",
                        "example": "user or admin or company"
                    },
                    
                    "address_id": {
                        "title": "Address ID",
                        "description": "The ID of the address",
                        "type": "big integer",
                        "format": 99
                    },
                    "membership_id": {
                        "title": "Membership ID ",
                        "description": "This is the ID of the membership that the user or organization choose",
                        "type": "big integer",
                        "example": 8
                    },
                    "email_verified_at": {
                        "title": "Email verified at",
                        "description": "Email verified at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "rememberToken": {
                        "title": "Login Token",
                        "description": "The token optained when user logged in",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                }
            },
...
        }
    }
}